Station and StationMarket Initial Work
Started Station object and StationMarket class implementation.
Added arrayCopy function to clone/copy arrays. Needed due to needing to
copy/clone an array of objects
StationMarketList is the list that holds all the stations implemented in
the game. Used as a master list to populate the StationMarket deck.

Station : The struct for a station card

StationMarket : Class that performs the functions of the station market.
This includes creating the deck of stations and manipulating the deck,
such as drawing cards and populating the deck.