Added marketCheck()
marketCheck() function in StationMarket class added. Function will check
the number of stations in the current and future market and will fill it
back up to the standard amount (currently 4 in current and 4 in future).
The added card will be sorted based on price to amongst other cards in
the market.