Fixed errors,started Phase 3 works for StationsMarket class
marketCheck() function was not complete. It did not populate the current
and future market after refilling the market.

Added add2P3Market function which will take the highest costing card in
the future market and add it into the phase 3 deck, then performs
marketCheck

Added a marketCheck at the end of removeStationFromCurrentMarket
function to refill the market's cards.