Various works for Phase 3 implementation
Added setPhase function into Game class. Function will set the current
phase of the game to the number it is given.

Changed deckCount function in StationMarket class to stationDeckCount to
differentiate the deck from the phase 3 market deck.

Added p3MarketDeckCount in StationMarket class function which returns
the number of cards in the phase 3 market deck.

Added p3Transition function in StationMarket class which performs the
tasks needed to be done when transitioning to phase 3. It also sets the
current phase of the game to phase 3.

Added checks of which phase the game is in in marketCheck function in
StationMarket class. Also added the algorithm of market refilling which
is done in phase 3. marketCheck also now will trigger the transition to
phase 3 when the deck count reaches zero, i.e. drawing the phase 3 card
from the deck.